+++ src/pages/index.tsx
@@ -1,0 +1,35 @@
+import Head from 'next/head'
+import {Button} from "@mui/material";
+
+export default function Home() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-screen py-2">
+      <Head>
+        <title>Material UI test</title>
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className="flex flex-col items-center justify-center w-full flex-1 px-20 text-center">
+        <h1 className="text-6xl font-bold mb-8">
+          Material UI test
+        </h1>
+
+        <p>
+          <Button className="bg-green-500 hover:bg-green-400 text-white p-4 rounded-full">Material Button</Button>
+        </p>
+      </main>
+
+      <footer className="flex items-center justify-center w-full h-24 border-t">
+        <a
+          className="flex items-center justify-center"
+          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Powered by{' '}
+          <img src="/vercel.svg" alt="Vercel Logo" className="h-4 ml-2" />
+        </a>
+      </footer>
+    </div>
+  )
+}
